+++ CHANGELOG.md
@@ -4,11 +4,20 @@ All the changes made to runa are documented here.
 
 ## Unreleased
 
+### Added
+- Move function: New move function, by default mapped to `m`, moves files to a directory via input. [(#25)](https://github.com/alexm-dev/runa/pull/25)
+- Move auto-complete: New move function input has auto-complete, to find easier paths easier to move files into. [(#25)](https://github.com/alexm-dev/runa/pull/25)
+- Dialog separators: Added horizontal separators to Delete, Move and Find widgets/dialogs, to visually separate input from the rest of the widget.
+- Move to trash: Toggle to enable files to be moved to trash instead of permamt deletion.
+
 ### Fixed
 - Preview width truncation: Now correclty requests a inital preview based on the actual pane width of the preview pane
 
 #### Changed
 - Layout metrics debounce for preview: Added a debounce for layout metrics changing, resulting in less preview requests for chaning the terminal size.
+
+
+---
 
 
 ## [0.5.3] - 2026-01-17